+++ macros/generics/activeAuraOnly.js
@@ -1,0 +1,11 @@
+if (!game.modules.get("advanced-macros")?.active) {
+  ui.notifications.error("Advanced Macros is not enabled");
+  return;
+} else if(!game.modules.get("ActiveAuras")?.active) {
+  ui.notifications.error("ActiveAuras is not enabled");
+  return;
+}
+
+if (args[0].tag === "OnUse") {
+  return await AAhelpers.applyTemplate(args);
+}