+++ src/effects/specialSpells.js
@@ -93,6 +93,7 @@ import { resistanceEffect } from "./spells/resistance.js";
 import { shieldEffect } from "./spells/shield.js";
 import { shieldofFaithEffect } from "./spells/shieldofFaith.js";
 import { shillelaghEffect } from "./spells/shillelagh.js";
+import { silenceEffect } from "./spells/silence.js";
 import { sleepEffect } from "./spells/sleep.js";
 import { slowEffect } from "./spells/slow.js";
 import { spiderClimbEffect } from "./spells/spiderClimb.js";
@@ -547,6 +548,10 @@ export async function spellEffectAdjustment(document) {
     }
     case "Shillelagh": {
       document = await shillelaghEffect(document);
+      break;
+    }
+    case "Silence": {
+      document = await silenceEffect(document);
       break;
     }
     case "Sleep": {